Feature: In-product support for Cloud Trial (#7907)
