+++ packages/mattermost-redux/src/types/cloud.ts
@@ -22,6 +22,8 @@ export type Subscription = {
     seats: number;
     is_paid_tier: string;
     last_invoice?: Invoice;
+    trial_end_at: number;
+    is_free_trial: string;
 }
 
 export type Product = {
