+++ sass/base/_compass-icons.scss
@@ -117,6 +117,7 @@
 .icon-chart-bar:before { content: '\f128'; } /* '' */
 .icon-chart-line:before { content: '\f12a'; } /* '' */
 .icon-check:before { content: '\f12c'; } /* '' */
+.icon-check-outline-circle:before { content: '\f05e1'; } /* '' */
 .icon-checkbox-blank-outline:before { content: '\f131'; } /* '' */
 .icon-checkbox-marked:before { content: '\f132'; } /* '' */
 .icon-chevron-down:before { content: '\f140'; } /* '' */
